+++ tests/test-all.R
@@ -2,5 +2,10 @@
 library(testthat)
 library(REDCapR)
 
+Sys.setenv("redcapr_test_server" = "dev-2")
+# Sys.setenv("redcapr_test_server" = "bbmc")
+
+message("Using test server '", Sys.getenv("redcapr_test_server"), "'.")
+
 # source("R/helpers-testing.R")
 testthat::test_check("REDCapR")